inserted "start-cook" to human-data
Added pickle file with the same data as `clean_train_trials.pkl` and `clean_test_trials.pkl` human data, but with the additional "interact" to start cooking in new dynamics. Also the data was reformatted to be an easier-to-read python dictionary format. 

Also included the file to perform the conversion, and a file to test the conversion.